requirements.txt

tracdap-runtime == 0.9.0

pandas == 2.2.3
numpy <= 2.0.0
scipy == 1.15.1
matplotlib == 3.10.0
polars == 1.22.0
